Struct ropey::Rope

pub struct Rope { /* fields omitted */ }

A utf8 text rope.

Rope's atomic unit of data is Unicode scalar values (or chars in Rust). Except where otherwise documented, all methods that index into a rope or return an index into a rope do so by char index. This makes the API's intuitive and prevents accidentally creating a Rope with invalid utf8 data.

The primary editing operations available for Rope are insertion of text, deletion of text, splitting a Rope in two, and appending one Rope to another. For example:

let mut rope = Rope::from_str("Hello みんなさん!");
rope.remove(6..11);
rope.insert(6, "world");

assert_eq!(rope, "Hello world!");

Cloning Rope's is extremely cheap, taking only a few instructions and 8 bytes of memory, regardless of text size. This is accomplished by data sharing between Rope clones, and the memory used by clones only grows incrementally as the their contents diverge due to edits. All of this is thread safe, and clones can be sent freely between threads.

Rope tracks line endings and has efficient API's for working with lines. You can convert between char and line index, determining which line a given char is on or the char index of the beginning of a line:

let rope = Rope::from_str("Hello individual!\nHow are you?\nThis text has multiple lines!");

assert_eq!(rope.char_to_line(5), 0);
assert_eq!(rope.char_to_line(21), 1);

assert_eq!(rope.line_to_char(0), 0);
assert_eq!(rope.line_to_char(1), 18);
assert_eq!(rope.line_to_char(2), 31);

Rope is written to be fast and memory efficient. Except where otherwise documented, all editing and query operations execute in worst-case O(log N) time in the length of the rope. It is designed to work efficiently even for huge (in the gigabytes) and pathological (all on one line) texts. It should be able to handle just about anything you can throw at it.

Methods

impl Rope

pub fn new() -> Self

Creates an empty Rope.

pub fn from_str(text: &str) -> Self

Creates a Rope from a string slice.

Runs in O(N) time.

pub fn from_reader<T: Read>(reader: T) -> Result<Self>

Creates a Rope from the output of a reader.

Runs in O(N) time.

Errors

If the reader returns an error, from_reader stops and returns that error.

If non-utf8 data is encountered, an IO error with kind InvalidData is returned.

Note: some data from the reader is likely consumed even if there is an error.

pub fn len_bytes(&self) -> usize

Total number of bytes in the Rope.

Runs in O(1) time.

pub fn len_chars(&self) -> usize

Total number of chars in the Rope.

Runs in O(1) time.

pub fn len_lines(&self) -> usize

Total number of lines in the Rope.

Runs in O(1) time.

pub fn capacity(&self) -> usize

Total size of the Rope's text buffer space, in bytes.

This includes unoccupied text buffer space. You can calculate the unoccupied space with capacity() - len_bytes(). In general, there will always be some unoccupied buffer space.

Runs in O(N) time.

pub fn shrink_to_fit(&mut self)

Shrinks the Rope's capacity to the minimum possible.

This will rarely result in capacity() == len_bytes(). Rope stores text in a sequence of fixed-capacity chunks, so an exact fit only happens for texts that are both a precise multiple of that capacity and have code point boundaries that line up exactly with the capacity boundaries.

After calling this, the difference between capacity() and len_bytes() is typically under 1KB per megabyte of text in the Rope.

NOTE: calling this on a Rope clone causes it to stop sharing all data with its other clones. In such cases you will very likely be increasing total memory usage despite shrinking the Rope's capacity.

Runs in O(N) time, and uses O(log N) additional space during shrinking.

pub fn insert(&mut self, char_idx: usize, text: &str)

Inserts text at char index char_idx.

Runs in O(M + log N) time, where N is the length of the Rope and M is the length of text.

Panics

Panics if char_idx is out of bounds (i.e. char_idx > len_chars()).

pub fn insert_char(&mut self, char_idx: usize, ch: char)

Inserts a single char ch at char index char_idx.

Runs in O(log N) time, where N is the length of the Rope.

Panics

Panics if char_idx is out of bounds (i.e. char_idx > len_chars()).

pub fn remove<R: CharIdxRange>(&mut self, range: R)

Removes the text in the given char index range.

Uses range syntax, e.g. 2..7, 2.., etc. The range is in char indices.

Example

let mut rope = Rope::from_str("Hello world!");
rope.remove(5..);

assert_eq!("Hello", rope);

Panics

Panics if the start of the range is greater than the end, or if the end is out of bounds (i.e. end > len_chars()).

pub fn split_off(&mut self, char_idx: usize) -> Self

Splits the Rope at char_idx, returning the right part of the split.

Panics

Panics if char_idx is out of bounds (i.e. char_idx > len_chars()).

pub fn append(&mut self, other: Self)

Appends a Rope to the end of this one, consuming the other Rope.

pub fn byte_to_char(&self, byte_idx: usize) -> usize

Returns the char index of the given byte.

Panics

Panics if byte_idx is out of bounds (i.e. byte_idx > len_bytes()).

pub fn char_to_byte(&self, char_idx: usize) -> usize

Returns the byte index of the given char.

Panics

Panics if char_idx is out of bounds (i.e. char_idx > len_chars()).

pub fn char_to_line(&self, char_idx: usize) -> usize

Returns the line index of the given char.

Notes:

Lines are zero-indexed.

If char_idx is one-past-the-end, then one-past-the-end line index is returned. This is mainly unintuitive for empty ropes, which will return a line index of 1 for a char_idx of zero. Otherwise it behaves as expected.

Panics

Panics if char_idx is out of bounds (i.e. char_idx > len_chars()).

pub fn line_to_char(&self, line_idx: usize) -> usize

Returns the char index of the start of the given line.

Notes:

Lines are zero-indexed.

line_idx can be one-past-the-end, which will return one-past-the-end char index.

Panics

Panics if line_idx is out of bounds (i.e. line_idx > len_lines()).

pub fn char(&self, char_idx: usize) -> char

Returns the char at char_idx.

Panics

Panics if char_idx is out of bounds (i.e. char_idx >= len_chars()).

pub fn line(&self, line_idx: usize) -> RopeSlice

Returns the line at line_idx.

Note: lines are zero-indexed.

Panics

Panics if line_idx is out of bounds (i.e. line_idx >= len_lines()).

pub fn slice<R: CharIdxRange>(&self, range: R) -> RopeSlice

Gets an immutable slice of the Rope.

Uses range syntax, e.g. 2..7, 2.., etc. The range is in char indices.

Example

let rope = Rope::from_str("Hello world!");
let slice = rope.slice(..5);

assert_eq!("Hello", slice);

Panics

Panics if the start of the range is greater than the end, or if the end is out of bounds (i.e. end > len_chars()).

Important traits for Bytes<'a>
impl<'a> Iterator for Bytes<'a> type Item = u8;

pub fn bytes(&self) -> Bytes

Creates an iterator over the bytes of the Rope.

Important traits for Chars<'a>
impl<'a> Iterator for Chars<'a> type Item = char;

pub fn chars(&self) -> Chars

Creates an iterator over the chars of the Rope.

Important traits for Lines<'a>
impl<'a> Iterator for Lines<'a> type Item = RopeSlice<'a>;

pub fn lines(&self) -> Lines

Creates an iterator over the lines of the Rope.

Important traits for Chunks<'a>
impl<'a> Iterator for Chunks<'a> type Item = &'a str;

pub fn chunks(&self) -> Chunks

Creates an iterator over the chunks of the Rope.

pub fn to_string(&self) -> String

Returns the entire text of the Rope as a newly allocated String.