Struct ropey::RopeSlice

pub struct RopeSlice<'a> { /* fields omitted */ }

An immutable view into part of a Rope.

Methods

impl<'a> RopeSlice<'a>

pub fn len_bytes(&self) -> usize

Total number of bytes in the RopeSlice.

pub fn len_chars(&self) -> usize

Total number of chars in the RopeSlice.

pub fn len_lines(&self) -> usize

Total number of lines in the RopeSlice.

pub fn byte_to_char(&self, byte_idx: usize) -> usize

Returns the char index of the given byte.

Panics

Panics if byte_idx is out of bounds (i.e. byte_idx > len_bytes()).

pub fn char_to_byte(&self, char_idx: usize) -> usize

Returns the byte index of the given char.

Panics

Panics if char_idx is out of bounds (i.e. char_idx > len_chars()).

pub fn char_to_line(&self, char_idx: usize) -> usize

Returns the line index of the given char.

Panics

Panics if char_idx is out of bounds (i.e. char_idx > len_chars()).

pub fn line_to_char(&self, line_idx: usize) -> usize

Returns the char index of the start of the given line.

Note: lines are zero-indexed.

Panics

Panics if line_idx is out of bounds (i.e. line_idx > len_lines()).

pub fn char(&self, char_idx: usize) -> char

Returns the char at char_idx.

Panics

Panics if char_idx is out of bounds (i.e. char_idx >= len_chars()).

pub fn line(&self, line_idx: usize) -> RopeSlice<'a>

Returns the line at line_idx.

Note: lines are zero-indexed.

Panics

Panics if line_idx is out of bounds (i.e. line_idx >= len_lines()).

pub fn slice<R: CharIdxRange>(&self, range: R) -> Self

Returns a sub-slice of the RopeSlice in the given char index range.

Uses range syntax, e.g. 2..7, 2.., etc. The range is in char indices.

Panics

Panics if the start of the range is greater than the end, or the end is out of bounds (i.e. end > len_chars()).

Important traits for Bytes<'a>
impl<'a> Iterator for Bytes<'a> type Item = u8;

pub fn bytes(&self) -> Bytes<'a>

Creates an iterator over the bytes of the RopeSlice.

Important traits for Chars<'a>
impl<'a> Iterator for Chars<'a> type Item = char;

pub fn chars(&self) -> Chars<'a>

Creates an iterator over the chars of the RopeSlice.

Important traits for Lines<'a>
impl<'a> Iterator for Lines<'a> type Item = RopeSlice<'a>;

pub fn lines(&self) -> Lines<'a>

Creates an iterator over the lines of the RopeSlice.

Important traits for Chunks<'a>
impl<'a> Iterator for Chunks<'a> type Item = &'a str;

pub fn chunks(&self) -> Chunks<'a>

Creates an iterator over the chunks of the RopeSlice.

pub fn to_string(&self) -> String

Returns the entire text of the RopeSlice as a newly allocated String.

pub fn to_rope(&self) -> Rope

Creates a new Rope from the contents of the RopeSlice.